Documentation of color chooser and multi-symbol semicolon behaviour
I added a new paragraph describing the color chooser. This is necessary because I refer to it in the response letter.
I also added documentation regarding behaviour of finding symbols, when the symbol string contains semicolos.
Jayson, could you please check if my description of these two points is correct, and if not fix it accordingly? I tried to describe as best as I could but am not 100% sure that my descrription is correct.